requirements.txt
pyTelegramBotAPI==4.15.4
google-generativeai==0.8.4
python-dotenv==1.0.1
